custom annimated nav bar added
